requirements.txt
psycopg2
pymongo
pandas
python-dotenv
black
